Feature/export file

+++ PIMTool.Core/Models/Request/Base/BaseSearchRequest.cs
@@ -5,8 +5,8 @@ namespace PIMTool.Core.Models.Request.Base;
 public class BaseSearchRequest
 {
     [Required]
-    [Range(1, long.MaxValue)]
-    public long PageIndex { get; set; }
+    [Range(1, int.MaxValue)]
+    public int PageIndex { get; set; }
     [Required]
     [Range(1, int.MaxValue)]
     public int PageSize { get; set; }
